requirements.txt
asgiref==3.5.0
azure-core==1.22.1
azure-storage-blob==12.9.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
cryptography==36.0.1
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-embed-video==1.4.1
django-heroku==0.3.1
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
install==1.3.5
isodate==0.6.1
msrest==0.6.21
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.0
psycopg2==2.9.3
pycparser==2.21
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.0.1
urllib3==1.26.8
whitenoise==6.0.0
